Atualização dos endpoint e criação da estrutura de testes e pesquisas

+++ src/main/java/br/com/linx/cardValidator/repository/BinTypeRepository.java
@@ -1,12 +1,16 @@
 package br.com.linx.cardValidator.repository;
 
+import br.com.linx.cardValidator.model.Bin;
 import br.com.linx.cardValidator.model.BinType;
+import br.com.linx.cardValidator.model.Brand;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface BinTypeRepository extends JpaRepository<BinType, Integer> {
+public interface BinTypeRepository extends JpaRepository<BinType, Long> {
 
-    List<BinType> findByIdType(Long id_type);
+   List<BinType> findByIdType(Long id_type);
+
+
 
 }
